Split cross-domain controllers into proper domains
Split game.py:
- Game detail view stays in games/_detail.py
- Preference/played endpoints move to player/_preferences.py

Split event_player.py:
- Game browsing/filtering stays in games/_browse.py (renamed EventGamesController)
- Session planner moves to player/_planner.py (new PlannerController)

Phase 3b — controller splits complete.
