+++ convergence_games/domain/games/controllers/__init__.py
@@ -1,11 +1,11 @@
-from ._browse import EventPlayerController
+from ._browse import EventGamesController
 from ._components import MiscComponentsController
 from ._detail import GameController
 from ._search import SearchController
 from ._submit import SubmitGameController
 
 __all__ = [
-    "EventPlayerController",
+    "EventGamesController",
     "GameController",
     "MiscComponentsController",
     "SearchController",
