+++ convergence_games/domain/player/controllers/__init__.py
@@ -1,3 +1,5 @@
 from ._party import PartyController
+from ._planner import PlannerController
+from ._preferences import PreferencesController
 
-__all__ = ["PartyController"]
+__all__ = ["PartyController", "PlannerController", "PreferencesController"]
